Add CloudWatch alert for Redis database usage
We are adding two alerts for CloudWatch: `redis-elasticache-high-db-memory-warning`
and `redis-elasticache-high-db-memory-critical`.
`redis-elasticache-high-db-memory-warning` creates an alert at the warning
level when Redis memory usage reaches or exceeds the threshold value of
`${var.warning_database_usage_threshold}`%. Currently, this value is set to 90.
`redis-elasticache-high-db-memory-critical` creates an alert at the critical
level when Redis memory usage reaches or exceeds the threshold value of
`${var.critical_database_usage_threshold}`%. Currently, this value is set to 96.

[INSLY-2356][1]

[1]:https://app.clickup.com/t/2454960/INSLY-2356