Add the `alerts_enabled` variable
A variable called `alerts_enabled` was added to control the
creation of resources. We check the value of the `alerts_enabled`
variable. If it is equal to `false`, then `count` is set to `0`, which
means that no resource instances will be created.
If the value of `alerts_enabled` is not equal to false,
then `count` is set to the value of the `redis_clusters` variable.
Therefore, the number of resource instances created will depend on
the value of the `redis_clusters` variable.

By default, the variable is set to false.

To change the variables in the module, they need to be defined.
Example:
```
module "redis-prj-beta" {
  source = "github.com/x-qdo/tf_aws_elasticache_redis.git?ref=xxxxx"

  ...
  alerts_enabled                    = true
  warning_database_usage_threshold  = 90
  critical_database_usage_threshold = 96
  sns_topic_arn                     = sns_arn
```